requirements.txt
numpy
pandas
python-dotenv
pyodbc
scikit-learn
#-e .